+++ pkg/scad/boolean/intersection.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package booleans
+package boolean
 
 import "github.com/micahkemp/scad/pkg/scad"
